requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
torch==1.12.1
pillow==12.0.0
timm==1.0.20
python-multipart==0.0.21
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-asgi==0.59b0